Makefile.am: use libzfcphbaapi.la as library name
After issuing make command during build process for standalone case
an error occurred:
...
libtool: link: gcc -DVLIBPATH=\"/usr/local/lib\" -D_REENTRANT -g -O2 -o zfcp_ping zfcp_ping.o  -lzfcphbaapi
/usr/bin/ld: cannot find -lzfcphbaapi
...

The reason for this error is that when using `-lzfcphbaapi` as
argument in `zfcp_ping_LDADD` and `zfcp_show_LDADD` automake is not
generating an explicit build dependency between the build of
`zfcp_ping`/`zfcp_show` and the build of `libzfcphbaapi.la`. Without
this dependency, Make is free to build the tools before the library is
build, and thus the build fails with the error shown above.

To fix this we use the library name `libzfcphbaapi.la` as argument and
so signal to automake, that `zfcp_ping`/`zfcp_show` have a direct
dependency on it. With this automake will create an explicit
dependency in the resulting Makefile, and Make will have to order the
build of the tools to after the library is build.

Signed-off-by: Fedor Loshakov <loshakov@linux.ibm.com>
Reviewed-by: Steffen Maier <maier@linux.ibm.com>
Reviewed-by: Benjamin Block <bblock@linux.ibm.com>
Signed-off-by: Steffen Maier <maier@linux.ibm.com>