vlib_sg_io: fix array overrun on converting 3-byte to integer
Fix by individually accessing only the 3 valid array entries
and manually calculate the resulting integer value by interpreting the
source as big-endian. But since the array is 3 -byte and the resulting
integer value is stored in a uint32_t type variable. Since integer
variable is uint32_t the first byte will be '0' and the remaining value
will be the 3-byte port_id.
So in the caller function sg_io_sendRNID() where the 'd_id' received
from getDidFromWWN() need to be left shifted by 8 bits before copying
it to the 'port_id' which is the part of cdb.

Signed-off-by: Steffen Maier <maier@linux.ibm.com>
Signed-off-by: M Nikhil <nikh1092@linux.ibm.com>
Acked-by: Fedor Loshakov <loshakov@linux.ibm.com>
Signed-off-by: Nihar Panda <niharp@linux.ibm.com>