vlib_sysfs: prevent SerialNumber field from overflow
The attr string variable has 80 characters maximum length and is used to
contain different property values from sfhelper_getProperty(). During
copy of serial number property from attr to SerialNumber variable with
strcpy() there is a potential string overflow, because maximum length of
SerialNumber string is 64. Avoid this situation by using strncpy() to
copy the SerialNumber from attr and the number of bytes here is the
minimum value of the lengths "sizeof(SerialNumber)" and "strlen(attr)".
strlcpy() is replaced with strncpy() as strlcpy() is not available on
the RHEL 9.4 distro.

[nikh1092@linux.ibm.com: edited the previous commit message to use
strncpy() instead of strlcpy() and specifying the number of bytes
to be copied is the minimum value of the strlen of 'attr' and the
size of SerialNumber]
Signed-off-by: M Nikhil <nikh1092@linux.ibm.com>
Acked-by: Fedor Loshakov <loshakov@linux.ibm.com>
Signed-off-by: Nihar Panda <niharp@linux.ibm.com>